--- BackEnd/src/main/kotlin/quest/marketstack/TradingApp/TradingAppApplication.kt
@@ -1,11 +1,0 @@
-package quest.marketstack.TradingApp
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class TradingAppApplication
-
-fun main(args: Array<String>) {
-	runApplication<TradingAppApplication>(*args)
-}
