--- BackEnd/src/main/kotlin/quest/marketstack/TradingApp/datasource/OpenTrades/MongoOpenTradeRepository.kt
@@ -1,10 +1,0 @@
-package quest.marketstack.TradingApp.datasource.OpenTrades
-
-import org.springframework.data.mongodb.repository.MongoRepository
-import org.springframework.stereotype.Repository
-import quest.marketstack.TradingApp.model.OpenTrade
-
-@Repository
-interface MongoOpenTradeRepository : MongoRepository<OpenTrade, String> {
-
-}
