--- BackEnd/src/main/kotlin/quest/marketstack/TradingApp/service/OpenTrades/OpenTradeServiceInterface.kt
@@ -1,14 +1,0 @@
-package quest.marketstack.TradingApp.service.OpenTrades
-
-import quest.marketstack.TradingApp.datasource.ClosedTrades.ClosedTradeDataSource
-import quest.marketstack.TradingApp.model.ClosedTrade
-import quest.marketstack.TradingApp.model.OpenTrade
-import quest.marketstack.TradingApp.model.TradeExec
-
-interface OpenTradeServiceInterface {
-    fun getTrades(): Collection<OpenTrade>
-    fun getTrade(id: String): OpenTrade?
-    fun addTradeExecs(execList: Collection<TradeExec>): Collection<TradeExec>
-    fun closeTrades(closedDataSource: ClosedTradeDataSource): Collection<ClosedTrade>
-
-}
