+++ BackEnd/src/main/kotlin/quest/marketstack/tradingapp/TradingAppApplication.kt
@@ -1,0 +1,11 @@
+package quest.marketstack.tradingapp
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class TradingAppApplication
+
+fun main(args: Array<String>) {
+    runApplication<TradingAppApplication>(*args)
+}
