+++ BackEnd/src/main/kotlin/quest/marketstack/tradingapp/domain/closedtrades/ClosedTradeDataSource.kt
@@ -1,0 +1,13 @@
+package quest.marketstack.tradingapp.domain.closedtrades
+
+import org.springframework.stereotype.Repository
+import quest.marketstack.tradingapp.model.ClosedTrade
+
+@Repository
+interface ClosedTradeDataSource {
+    fun retrieveTrades(): Collection<ClosedTrade>
+
+    fun retrieveTrade(id: String): ClosedTrade?
+
+    fun createTrade(execList: Collection<ClosedTrade>): Collection<ClosedTrade>
+}
