+++ BackEnd/src/main/kotlin/quest/marketstack/tradingapp/service/closedtrades/ClosedTradeServiceInterface.kt
@@ -1,0 +1,9 @@
+package quest.marketstack.tradingapp.service.closedtrades
+
+import quest.marketstack.tradingapp.model.ClosedTrade
+
+interface ClosedTradeServiceInterface {
+    fun getTrades(): Collection<ClosedTrade>
+
+    fun getTrade(id: String): ClosedTrade?
+}
