--- BackEnd/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.application.name=TradingApp
-spring.data.mongodb.uri=mongodb://localhost:27017/MarketStack
-spring.profiles.active=default
