--- BackEnd/src/test/kotlin/quest/marketstack/TradingApp/ClosedTrades/mock/MockClosedOpenTradeDataSourceTest.kt
@@ -1,17 +1,0 @@
-package quest.marketstack.TradingApp.ClosedTrades.mock
-
-import quest.marketstack.TradingApp.datasource.mock.MockClosedTradeDataSource
-import org.assertj.core.api.Assertions
-import org.junit.jupiter.api.Assertions.*
-import org.junit.jupiter.api.Test
-
-class MockClosedOpenTradeDataSourceTest{
-    private val mockDataSource = MockClosedTradeDataSource()
-
-    @Test
-    fun `should provide all closed trades in database`() {
-        val mockTrades = mockDataSource.retrieveTrades()
-
-        assertNotNull(mockTrades)
-    }
-}
