--- BackEnd/src/test/kotlin/quest/marketstack/TradingApp/OpenTrades/TradingAppApplicationTests.kt
@@ -1,13 +1,0 @@
-package quest.marketstack.TradingApp.OpenTrades
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class TradingAppApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
