--- BackEnd/src/test/kotlin/quest/marketstack/TradingApp/OpenTrades/service/OpenOpenTradeServiceTest.kt
@@ -1,23 +1,0 @@
-package quest.marketstack.TradingApp.OpenTrades.service
-
-import io.mockk.mockk
-import io.mockk.verify
-import org.junit.jupiter.api.Test
-import quest.marketstack.TradingApp.datasource.ClosedTrades.ClosedTradeDataSource
-import quest.marketstack.TradingApp.datasource.OpenTrades.OpenTradeDataSource
-import quest.marketstack.TradingApp.service.OpenTrades.OpenTradeService
-
-private val dataSource: OpenTradeDataSource = mockk(relaxed = true)
-private val closedDataSource: ClosedTradeDataSource = mockk(relaxed = true)
-
-private val tradeService = OpenTradeService(dataSource, closedDataSource)
-
-class OpenOpenTradeServiceTest{
-    @Test
-    fun `should call it's data source to retrieve open trades`() {
-
-        val trades = tradeService.getTrades()
-
-        verify(exactly = 1){ dataSource.retrieveTrades()}
-    }
-}
