+++ BackEnd/src/test/kotlin/quest/marketstack/tradingapp/opentrades/TradingAppApplicationTests.kt
@@ -1,0 +1,11 @@
+package quest.marketstack.tradingapp.opentrades
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class TradingAppApplicationTests {
+    @Test
+    fun contextLoads() {
+    }
+}
