+++ BackEnd/src/test/resources/application-test.yml
@@ -1,0 +1,15 @@
+spring:
+  application:
+    name: TradingApp
+  data:
+    mongodb:
+      uri: mongodb://localhost:27017/MarketStack
+
+api:
+  closed-trade:
+    base-url: /api/v1/closedTrades
+    collection: Trade-Closed
+  open-trade:
+    base-url: /api/v1/openTrades
+    collection: Trade-Open
+
